Adds GooglePreferencesFragment Adds PreferencesFragment Adds OtherPreferencesFragment
