+++ collect_app/src/main/java/org/odk/collect/android/preferences/GooglePreferencesActivity.java
@@ -14,23 +14,8 @@
 
 package org.odk.collect.android.preferences;
 
-import android.accounts.Account;
-import android.accounts.AccountManager;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.EditTextPreference;
-import android.preference.ListPreference;
-import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
 import android.preference.PreferenceActivity;
-import android.preference.PreferenceCategory;
-import android.text.InputFilter;
-import android.widget.Toast;
-
-import org.odk.collect.android.R;
-import org.odk.collect.android.utilities.UrlUtils;
-
-import java.util.ArrayList;
 
 /**
  * Handles Google specific preferences.
@@ -39,93 +24,10 @@ import java.util.ArrayList;
  */
 public class GooglePreferencesActivity extends PreferenceActivity {
 
-    private ListPreference mSelectedGoogleAccountPreference;
-    protected EditTextPreference mGoogleSheetsUrlPreference;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.google_preferences);
-
-        boolean adminMode = getIntent().getBooleanExtra(PreferencesActivity.INTENT_KEY_ADMIN_MODE,
-                false);
-
-        SharedPreferences adminPreferences = getSharedPreferences(
-                AdminPreferencesActivity.ADMIN_PREFERENCES, 0);
-
-        mSelectedGoogleAccountPreference = (ListPreference) findPreference(
-                PreferenceKeys.KEY_SELECTED_GOOGLE_ACCOUNT);
-        PreferenceCategory googlePreferences = (PreferenceCategory) findPreference(
-                getString(R.string.google_preferences));
-
-        mSelectedGoogleAccountPreference
-                .setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-                    @Override
-                    public boolean onPreferenceChange(Preference preference, Object newValue) {
-                        int index = ((ListPreference) preference).findIndexOfValue(newValue
-                                .toString());
-                        String value =
-                                (String) ((ListPreference) preference).getEntryValues()[index];
-                        ((ListPreference) preference).setSummary(value);
-                        return true;
-                    }
-                });
-        mSelectedGoogleAccountPreference.setSummary(mSelectedGoogleAccountPreference.getValue());
-
-        boolean googleAccountAvailable = adminPreferences.getBoolean(
-                AdminKeys.KEY_CHANGE_GOOGLE_ACCOUNT, true);
-        if (!(googleAccountAvailable || adminMode)) {
-            googlePreferences.removePreference(mSelectedGoogleAccountPreference);
-        }
-
-        mGoogleSheetsUrlPreference = (EditTextPreference) findPreference(
-                PreferenceKeys.KEY_GOOGLE_SHEETS_URL);
-        mGoogleSheetsUrlPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object newValue) {
-                String url = newValue.toString();
-
-                // remove all trailing "/"s
-                while (url.endsWith("/")) {
-                    url = url.substring(0, url.length() - 1);
-                }
-
-                if (UrlUtils.isValidUrl(url) || url.length() == 0) {
-                    preference.setSummary(newValue.toString());
-                    return true;
-                } else {
-                    Toast.makeText(getApplicationContext(), R.string.url_error, Toast.LENGTH_SHORT)
-                            .show();
-                    return false;
-                }
-            }
-        });
-        mGoogleSheetsUrlPreference.setSummary(mGoogleSheetsUrlPreference.getText());
-        mGoogleSheetsUrlPreference.getEditText().setFilters(new InputFilter[]{
-                new ControlCharacterFilter(), new WhitespaceFilter()
-        });
-
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        // get list of google accounts
-        final Account[] accounts = AccountManager.get(getApplicationContext()).getAccountsByType(
-                "com.google");
-        ArrayList<String> accountEntries = new ArrayList<String>();
-        ArrayList<String> accountValues = new ArrayList<String>();
-
-        for (int i = 0; i < accounts.length; i++) {
-            accountEntries.add(accounts[i].name);
-            accountValues.add(accounts[i].name);
-        }
-        accountEntries.add(getString(R.string.no_account));
-        accountValues.add("");
-
-        mSelectedGoogleAccountPreference.setEntries(accountEntries
-                .toArray(new String[accountEntries.size()]));
-        mSelectedGoogleAccountPreference.setEntryValues(accountValues
-                .toArray(new String[accountValues.size()]));
+        getFragmentManager().beginTransaction().replace(android.R.id.content, new GooglePreferencesFragment()).commit();
     }
 }
