+++ collect_app/src/main/java/org/odk/collect/android/preferences/OtherPreferencesActivity.java
@@ -15,56 +15,24 @@
 package org.odk.collect.android.preferences;
 
 import android.os.Bundle;
-import android.preference.EditTextPreference;
-import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
-import android.text.InputFilter;
 
 import org.odk.collect.android.R;
+
+import java.util.List;
 
 /**
  * Handles 'other' specific preferences.
  *
  * @author Carl Hartung (chartung@nafundi.com)
  */
-public class OtherPreferencesActivity extends AggregatePreferencesActivity
-        implements OnPreferenceChangeListener {
-
-    protected EditTextPreference mSubmissionUrlPreference;
-    protected EditTextPreference mFormListUrlPreference;
+public class OtherPreferencesActivity extends PreferencesActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-        addPreferencesFromResource(R.xml.other_preferences);
-
-        mFormListUrlPreference = (EditTextPreference) findPreference(
-                PreferenceKeys.KEY_FORMLIST_URL);
-        mSubmissionUrlPreference = (EditTextPreference) findPreference(
-                PreferenceKeys.KEY_SUBMISSION_URL);
-
-        InputFilter[] filters = {new ControlCharacterFilter(), new WhitespaceFilter()};
-
-        mServerUrlPreference.getEditText().setFilters(filters);
-
-        mFormListUrlPreference.setOnPreferenceChangeListener(this);
-        mFormListUrlPreference.setSummary(mFormListUrlPreference.getText());
-        mFormListUrlPreference.getEditText().setFilters(filters);
-
-        mSubmissionUrlPreference.setOnPreferenceChangeListener(this);
-        mSubmissionUrlPreference.setSummary(mSubmissionUrlPreference.getText());
-        mSubmissionUrlPreference.getEditText().setFilters(filters);
+        getFragmentManager()
+                .beginTransaction()
+                .replace(android.R.id.content, new OtherPreferencesFragment())
+                .commit();
     }
-
-    /**
-     * Generic listener that sets the summary to the newly selected/entered
-     * value
-     */
-    @Override
-    public boolean onPreferenceChange(Preference preference, Object newValue) {
-        preference.setSummary((CharSequence) newValue);
-        return true;
-    }
-
 }
