+++ collect_app/src/main/java/org/odk/collect/android/preferences/OtherPreferencesFragment.java
@@ -1,0 +1,56 @@
+package org.odk.collect.android.preferences;
+
+import android.os.Bundle;
+import android.preference.EditTextPreference;
+import android.preference.Preference;
+import android.text.InputFilter;
+
+import org.odk.collect.android.R;
+
+/**
+ * Created by shobhit on 6/3/17.
+ */
+
+public class OtherPreferencesFragment extends AggregatePreferencesFragment implements Preference.OnPreferenceChangeListener {
+
+    protected EditTextPreference mSubmissionUrlPreference;
+    protected EditTextPreference mFormListUrlPreference;
+
+
+    @Override
+    public void onCreate(final Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.other_preferences);
+
+        mFormListUrlPreference = (EditTextPreference) findPreference(
+                PreferenceKeys.KEY_FORMLIST_URL);
+        mSubmissionUrlPreference = (EditTextPreference) findPreference(
+                PreferenceKeys.KEY_SUBMISSION_URL);
+
+        InputFilter[] filters = {new ControlCharacterFilter(), new WhitespaceFilter()};
+
+        mServerUrlPreference.getEditText().setFilters(filters);
+
+        mFormListUrlPreference.setOnPreferenceChangeListener(this);
+        mFormListUrlPreference.setSummary(mFormListUrlPreference.getText());
+        mFormListUrlPreference.getEditText().setFilters(filters);
+
+        mSubmissionUrlPreference.setOnPreferenceChangeListener(this);
+        mSubmissionUrlPreference.setSummary(mSubmissionUrlPreference.getText());
+        mSubmissionUrlPreference.getEditText().setFilters(filters);
+
+    }
+
+
+    /**
+     * Generic listener that sets the summary to the newly selected/entered
+     * value
+     */
+    @Override
+    public boolean onPreferenceChange(Preference preference, Object newValue) {
+        preference.setSummary((CharSequence) newValue);
+        return true;
+    }
+
+
+}
